requirements.txt
fastapi[standard]
python-dotenv
jinja2
google-genai
json-repair
httpx
chromadb
icalendar
recurring-ical-events
mutagen
faster-whisper
deepgram-sdk
